requirements.txt
neo4j
groq
fastapi
uvicorn
requests
tree-sitter>=0.22.0
tree-sitter-python
tree-sitter-typescript
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv
